+++ src/WebCompiler/Compile/LessCompiler.cs
@@ -116,6 +116,12 @@ namespace WebCompiler
             if (options.StrictMath)
                 arguments += " --strict-math=on";
 
+            if (!options.IECompat)
+                arguments += " --no-ie-compat";
+
+            if (options.StrictMath)
+                arguments += " --strict-math=on";
+
             if (options.StrictUnits)
                 arguments += " --strict-units=on";